move some classes to sft/utils.py
